\name{datasample}
\alias{datasample}
\docType{data}
\title{
A sample dataset for a community
}
\description{
A sample community matrix containing 8 plots and 14 species, from Gutianshan, Zhejiang, China.
}
\usage{data(datasample)}

\details{
Values are the importance value for each species.
}
\source{
Hu Zheng-hua, Qian Hai-Yuan, Yu Ming-jian. 2009. The niche of dominant species populations in Castanopsis eyrei forest in Gutian Mountain National Natural Reserve. Acta Ecologica Sinica. Vol.29, 3670-3677
}
\examples{
data(datasample)
datasample
}
\keyword{datasets}
