\name{testdata}
\alias{testdata}
\docType{data}
\title{
A sample dataset
s}
\description{
A sample dataset
}
\usage{data(testdata)}
\format{
  A dataframe with 11 observations on the following 3 variables.
  \describe{
    \item{\code{plotname}}{a factor with levels \code{plot1}, \code{plot2}, \code{plot3}.}
    \item{\code{species}}{a factor with levels \code{sp1} to \code{sp7}. }
    \item{\code{abundance}}{a numeric vector indicating number of individuals appeared in each plot.}
  }
}

\examples{
data(testdata)
testdata
}
\keyword{datasets}
