\name{XYname}
\alias{XYname}
\title{
Generating vector of XY labels
}
\description{
Generating vector of XY labels by providing number of rows and columns
}
\usage{
XYname(x, y)
}
\arguments{
  \item{x}{
  Number of X labels
}
  \item{y}{
  Number of Y labels
}
}

\value{
Vector of XY labels
}
\references{
None
}
\author{
Jinlong Zhang \email{jinlongzhang01@gmail.com}
}

\seealso{
\code{\link{lab.mat}} for converting the vector to matrix of XY labels.
}
\examples{

## XYname() example
XYname(4,6)

}

\keyword{ turnover }
