\name{datasample}
\alias{datasample}
\docType{data}
\title{
community matrix example
}
\description{
Community matrix
}
\usage{data(datasample)}
\format{
  A data frame with 8 plots on the following 14 species.
}
\details{
 including 14 species, 8 plots of Gutianshan Natural Reserve, Zhejiang, China, values are the value of importance in the plot for each species.
}
\source{
Hu Zheng-hua, Qian Hai-Yuan, Yu Ming-jian. 2009. The niche of dominant species populations in Castanopsis eyrei forest in Gutian Mountain National Natural Reserve. Acta Ecologica Sinica. Vol.29, 3670-3677
}
\examples{
data(datasample)
}
\keyword{datasets}
